requirements.txt
asgiref==3.8.1
Django==4.2.13
sqlparse==0.5.0
psycopg2-binary==2.9.9
pillow==10.3.0
djangorestframework==3.15.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.0.4
drf-yasg==1.21.7
django-cors-headers==4.3.1
requests==2.32.2
coverage==7.5.1